+++ src/types/actionlib_msgs.ts
@@ -1,4 +1,4 @@
-import { std_msgs } from "../types/std_msgs";
+import { std_msgs } from "./std_msgs.js";
 import { GoalStatus as GoalStatusEnum } from "../core/GoalStatus.js";
 
 export namespace actionlib_msgs {
@@ -12,7 +12,7 @@ export namespace actionlib_msgs {
     text?: string;
   }
   export interface GoalStatusArray {
-    header: std_msgs.Header;
+    header: std_msgs.ROS1Header;
     status_list: GoalStatus[];
   }
 }
